加入纹理的显示。

+++ Engine/SemperEngine/GWinText.cpp
@@ -17,7 +17,7 @@ namespace SemperEngine
 			size_t size = wstr.size();
 			size_t bufferSize = WideCharToMultiByte(CP_UTF8, WC_NO_BEST_FIT_CHARS, data, -1, nullptr, 0, NULL, NULL);
 			unique_ptr<char> buffer2(new char[bufferSize]);
-			WideCharToMultiByte(CP_UTF8, WC_NO_BEST_FIT_CHARS, data, -1, buffer2.get(), bufferSize, NULL, NULL);
+			WideCharToMultiByte(CP_UTF8, WC_NO_BEST_FIT_CHARS, data, -1, buffer2.get(), (int)bufferSize, NULL, NULL);
 			string result = buffer2.get();
 			return result;
 		}
@@ -30,7 +30,7 @@ namespace SemperEngine
 			size_t size = str.size();
 			size_t bufferSize = MultiByteToWideChar(CP_UTF8, WC_NO_BEST_FIT_CHARS, data, -1, nullptr, 0);
 			unique_ptr<wchar_t> buffer2(new wchar_t[bufferSize]);
-			MultiByteToWideChar(CP_UTF8, WC_NO_BEST_FIT_CHARS, data, -1, buffer2.get(), bufferSize);
+			MultiByteToWideChar(CP_UTF8, WC_NO_BEST_FIT_CHARS, data, -1, buffer2.get(), (int)bufferSize);
 			wstring result = buffer2.get();
 			return result;
 		}
